+++ challenge/client/src/components/Dashboard/ChartTable.js
@@ -9,6 +9,8 @@ import PersonPinIcon from '@material-ui/icons/PersonPin';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
 import ThisWeekChart from './ThisWeekChart';
+import OverallChart from './OverallChart';
+
 
 function TabPanel(props) {
   const { children, value, index, ...other } = props;
@@ -76,7 +78,7 @@ export default function ScrollableTabsButtonForce(props) {
         <ThisWeekChart data={props.data}/>
       </TabPanel>
       <TabPanel value={value} index={1}>
-        Item Two
+        <OverallChart data={props.overall}/>
       </TabPanel>
     </div>
   );
