+++ challenge/routes/api/users.js
@@ -1,11 +1,10 @@
 const router = require("express").Router();
-const challengeMeController = require("../../controllers/challengeMeController");
+const userController = require("../../controllers/userController")
 
 // Matches with "/api/books"
 router.route("/create")
-  .post(challengeMeController.createUser)
-
+      .post(userController.createUser)
 router.route("/find/:firebaseId")
-  .get(challengeMeController.findUser)
+      .get(userController.findUser)
 
 module.exports = router;
